requirements.txt
Django>=4.1.2,<4.2
djangorestframework>=3.14.0,<3.15
Markdown>=3.4.1,<3.5
django-filter>=22.1,<22.2
djangorestframework-simplejwt>=5.2.2,<5.3
